Add files via upload
This is a project in Data Analytics and Visualisation. where we were supposed to conduct Exploratory Data Analysis on the top universities of 2012-2015. I contextualised this project in the eyes of Charlie, a girl who is looking for a suitable university for her higher education. 

This project was done alone, and I was awarded the highest grade of distinction for this project.